Fixing compiler warnings for type mismatch in printf and fprintf functions. Issues are that Windows defines size_t as long long int and Linux defines size_t as long int. The best way is to just do a "C" style cast in the varargs section of the function to quiet the compiler.

+++ Source/Plugins/IO/IOFilters/AbaqusSurfaceMeshWriter.cpp
@@ -339,17 +339,17 @@ int32_t AbaqusSurfaceMeshWriter::writeFeatures(FILE* f)
       // Only print 15 Triangles per line
       if (lineCount == 15)
       {
-        fprintf (f, ", %lld\n", t);
+        fprintf (f, ", %lld\n", (long long int)(t));
         lineCount = 0;
       }
       else if(lineCount == 0) // First value on the line
       {
-        fprintf(f, "%lld", t);
+        fprintf(f, "%lld", (long long int)(t));
         lineCount++;
       }
       else
       {
-        fprintf(f, ", %lld", t);
+        fprintf(f, ", %lld", (long long int)(t));
         lineCount++;
       }
